+++ src/main/java/janggi/piece/Piece.java
@@ -2,7 +2,6 @@ package janggi.piece;
 
 import janggi.position.Path;
 import janggi.position.Position;
-import janggi.view.PieceType;
 
 public abstract class Piece {
